Deterministic proto message marshaling and hashing
Since we can't rely on the stability of the protobuf package's
deterministic marshaling, we need an implementation that we control.  In
support of that, make a slightly modified copy of the deterministic,
reflect based protobuf implementation in our protomsg package.

The protomsg package should only be used when an encoding must be stable
across time and space. In particular, this is required for transaction
ID as a hash of the encoding of elements within the transaction.

Note that, as of right now, the reflect based implementation and the
optimized, deterministic marshaling from the protobuf package generate
the same encoding. As long as this holds true, we can use the official
implementation for encoding as it performs an order of magnitude better
than the reflect based implementation.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>