Add basic BatikConfig parsing
When starting a batik instance, it will attempt to load a configuration
file from one of the following sources in order:
  1. TODO: command line flag for batik cfg path
  2. $BATIK_CFG_PATH
  3. $(pwd)/batik.yaml
  4. $XDG_CONFIG_HOME/batik/batik.yaml
  5. $HOME/.config/batik/batik.yaml
If no configuration file is found from a user specified file in 1. or 2.
it will return an error.
If no configuration file is found in any other location, configuration
must be passed either via environment variables or command line flags.

Configuration overrides take precedence as follows:
  1. Command line flags
  2. TODO: Env vars
  3. Configuration from yaml

Signed-off-by: Danny Cao <dcao@us.ibm.com>