Add config struct decoding
A decoder instance first attempts to parse env vars designated by the
parseTag tag on the config struct. If a value was already loaded from a
yaml config file, it will attempt to override with the associated env
var if it is set. If the value was not loaded from a config file and no
env var is set, it will attempt to load a default value designated by
the defaultTag tag on the config struct.

Signed-off-by: Danny Cao <dcao@us.ibm.com>