The tag for default is 'default' not 'example'
The 'example' will eventually be used for doc generation. It's role is
to provide a value to use as an example in documentation. The canonical
use case is required configuration elements where a default value does
not make a good example or where a default value does not make sense.

A concrete example from Fabric would be PeerID. That configuration value
is absolutely required, we would want to provide an example of the value
in documentation, but we want to ensure that a user provides an
unambiguous value at runtime.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>