Bump urfave/cli to pick up error processing changes
A pull request was merged on April 28 that treats errors from app.Before
the same as other errors. This behavior is desired as it prevents the
usage message from going to the app.Writer and writes the actual error
message to app.ErrWriter.

We'll want to pick up the formal release that includes this change when
it's available.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>