Remove log.NewWriter
The original call has been commented out, the function wasn't tested,
and added functionality to the logging package that is better
implemented by the logging consumer.

For example, in practice, writing to a log file requires additional
machinery for rotation, management, and signal processing. These can all
be implemented in a way that exposes a single writer instance to the
logging framework.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>