Create config in app, wire flags to config instance
Building on the configuration experiment, this creates an instance of
configuration defaults for the batik app. This instance is wired into
the commands such that the commands can acquire Flag for the
appropriate configuration element for processing.

A side effect of these changes is that the configuration is no longer
required as shared data on the context. Another side effect is that the
address flag on "start" was changed to "listen-address".

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>