Wrap flags to avoid overwriting parsed config
Flags and their initial values are setup when the app and commands are
instantiated but the flags and flag set instances are not created until
the command executes. Since the app reads the configuration file between
the setup and execution, this resulted in the default values from
instantiation overwriting the values read from the config file.

Given our use case of setting up defaults, reading config, and then
applying flags, this commit creates wrappers around each of the flag
types from the cli package to use updated values as flag defaults before
before the standard library is invoked to create the flag var.

It's a bit of a hack but low touch and relegated to the bit of the code
that creates the config structs and flags for the CLI.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>