Introduce pretty writer for colorizing log output
The pretty writer sits in front of the logger's writer and intercepts
logfmt lines to prettify based on some default options for colors. If
the color flag is set to yes or auto and tty is detected, the format
will be ignored and instead encoded as logfmt before being post
processed to a color formatted human readable log line.

Signed-off-by: Danny Cao <dcao@us.ibm.com>