Sync pretty writer to encoder config
* Pretty writer now uses the encoder config for determining which keys
to process for reserved fields
* A timeparser.TimeParser func is now passed to the pretty writer for
handling time parsing. For now this is just a ParseUnixTime but in the
future we could use this to parse times with different layouts
* The pretty writer now assumes the structure of the logfmt line, it
will always first attempt to parse time, level, name, caller, message,
and then remaining fields. If any reserved field is missing it will
currently throw off the parser.

TODO: The pretty writer should be resilient to parsing logfmt lines that
are missing certain reserved fields.

Signed-off-by: Danny Cao <dcao@us.ibm.com>