Move listen-address from server to server.grpc
The gRPC gateway is exposed over http instead of grpc. Given the
restrictions and performance implications from multiplexing gRPC and
HTTP on the same transport using either cmux or the HTTP handler
approach, we're going to use separate server endpoints for the
protocols.

In preparation, the gRPC endpoint is being moved off of "server" and
down to "server.grpc". A new "server.http" configuration will be used to
configure the http endpoint. TLS configuration will be shared across
both.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>