Cleanup and fix pretty.ParseUnixTime
There was code in ParseUnixTime that was checking the result of parsing
the fractional number of seconds elapsed since Unix epoch, converted to
an int64 (throwing away the frational bit) against various large
numbers. This implementation did not reflect the value that was being
parsed.

The implementation was changed to parse the value into two strings that
hold seconds and nanoseconds. The string containing seconds is parsed
as an integer and the string containing nanoseconds is parsed as a float
and scaled by 1e9. These values are then provided to time.Unix to create
the time stamp.

This resulted in a failure in the one of the variations of TestWrite so
I updated that and cleaned it up a bit.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>