Fix logging configuration in app/batik
The time encoder used in the production implementation was working
around the buggy implementation within the logging package. After fixing
the bug, the non-default time encoder needs to be removed.

While testing this change, I noticed that `--format json` was not being
honored when specified. This is related to the structure of the switch
that is used to construct the encoder. While the colorized encoder does
require a console, it should only be used when the logging format is
logfmt. This was addressed by reordering the cases in the switch
statement to handle the 'json' format first.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>