Make pretty.Writer adhere to io.Writer contract
The io.Writer contract indicates that Write should return the number of
bytes written *from* p and that return value should be in the range >= 0
and <= len(p). The existing implementation did not adhere to this
contract as it returned the number of bytes written to the embedded
writer. This value can be much larger than the size of p because of the
additional formatting that is added.

Instead of returning the number of bytes written, Write now returns the
length of the source on success and 0 on error.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>