Use Unsafe* variants of gRPC server interfaces
The gRPC bindings generated by the new plugin require service
implementations to embed an interface that satisfies the
"mutEmbedUnimplemented*" contract. As with most things gRPC and
protobuf, this is yet another parochial decision that assumes
compilation failures are a problem instead of a virtue.

This change moves from using the "Unimplemented" to "Unsafe" variants of
the mustEmedUnimplemented interface to help ensure compilation fails
when the implementation does not satisfy the service contract. Adding an
implementation that returns an "unimplemented" error is a trivial
exercise that makes the behavior explicit and evident to readers of the
code.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>