Remove contents of protos/tx/v1/resolved.proto
Resolved states are an artifact of validation where references are
"resolved" prior to invoking the validator. This concept should not
be present in the context of transaction submission API.

With the removal of these messages, the prototyping storev1 services
have been modified to embed state reference in the request and response
messages.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>