Pass input message length to wasm validator
Since we know the length of the message we're providing to the
validator, we can simply pass it along on invocation. This allows the
validator to allocate a buffer of the correct length and read the
request into it.

On the response side, there's no need to provide the length in the
stream since the length is an argument to write.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>