Sign digest of tx ID instead of the actual tx ID
Even though the transaction ID is a SHA256 merkle root digest of the
serialized transaction, some signature verification mechanisms
explicitly digest messages. This happens to be the case with the
ecdsa signature verification done in the rust signature crate.

Additionally, it appears it's good form to match the size of the digest
to the size of the signing key.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>