Create production tlscerts package
This commit extracts some of the cert generation function, specifically
what is needed to generate a self-signed server cert from the tested
package into a new tlscerts package.

It takes the approach of exposing a few template generation functions,
as well as a function to generate a CA.  Notably, it does not expose
functions for generating client or server certs issued by that CA, and
instead opts to leave them in the tested package.  If/when those
functions are required they can also be moved, but doing so now seemed
premature.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>