Add some tests to show flags override config files
We have a problem where global flags do not override the elements read
from the configuration file.

The current process does something like this:

1. Create a default configuration object
2. Register flags that update the configuration object
3. Start the cli.App with Run
4. The cli.App creates a flag.FlagSet, parses the flags, and sets the
   flag values.
5. The Before hook is called by Run and we get the value of the
   `--config` flag and then read the file.

Once we get to 5, the values from the configuration file overwrite what
was written when the flags were parsed.

For command flags (not global flags), the command line processing occurs
after the app's Before hook runs so they *do* overwrite the values read
from the configuration file.

And so, we have a bit of a chicken and egg problem. We want to take
advantage of the flag support from the cli but we need to read the
configuration file before the flags are processed.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>