IT enhanced to do wasm and non-wasm validation
This commit introduces a substancial performance hit in the integration
tests.  It adds about 20 seconds to the integration tests due to loading
the WASM validator in all tests.

We probably could/should consider only loading the WASM bits in the
tests that need it, but, it would have complicated the test structure,
perhaps requiring a 'JustBeforeEach' section to actually launch the
server, so leaving for the moment.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>