Move crate into 'rust' folder and rename it sigval
The rust code implements the signature based validator that's
implemented in go so call it 'sigval' for "signature validator" instead
of 'utxotx'. (The 'signature' create already exists and causes name
collisions.)

Also make some changes to avoid running a cargo build all the time and
change the build to generate a `release` module that's about 1/10 the
size of the `debug` version.

Instead of calling `cargo build` directly, the tests will run a make
target. TBD if this is better or worse...

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>