Move submit to namespace
In order to split submission and ordering as a first step, we need to
split submit and commit.  This first step moves the submit code to the
namespace package, and puts the committing half into its own structure.

Note, because we now put the tx onto the ledger before commit, and
because once ordering introduces an asynchronous delay it will not be
possible to prevent duplicate submission of the same tx receipt, this
change removes the duplicate tx submission checks and test.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>