Add support to store for commits
Since we need to store the tx for dissemination and ordering before
committment, we need another way to indicate that a particular txid has
committed other then the simple presence of a tx in the db.  This adds a
new keyspace for committed txids.  A txid can only commit once, and will
be associated with a particular local sequence and receipt.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>